Updated Booking Logic to validate start time and date

+++ Tikito/src/main/java/com/sunbeam/tikito/serviceimpl/BookingServiceImpl.java
@@ -1,9 +1,12 @@
 package com.sunbeam.tikito.serviceimpl;
 
+import java.time.LocalDate;
+import java.time.LocalTime;
 import java.util.ArrayList;
 import java.util.List;
 import java.util.Optional;
 
+import org.modelmapper.ModelMapper;
 import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.stereotype.Service;
 
@@ -12,6 +15,7 @@ import com.sunbeam.tikito.daos.BookingDao;
 import com.sunbeam.tikito.daos.SeatDao;
 import com.sunbeam.tikito.daos.ShowDao;
 import com.sunbeam.tikito.daos.UserDao;
+import com.sunbeam.tikito.dto.AllBookingsDto;
 import com.sunbeam.tikito.dto.BookedSeatDto;
 import com.sunbeam.tikito.dto.CancelTicketDto;
 import com.sunbeam.tikito.dto.TicketBookedDto;
@@ -22,7 +26,7 @@ import com.sunbeam.tikito.enums.BookingStatus;
 import com.sunbeam.tikito.enums.PaymentStatus;
 import com.sunbeam.tikito.exceptions.InvalidBookingException;
 import com.sunbeam.tikito.exceptions.InvalidSeatsException;
-import com.sunbeam.tikito.exceptions.ShowNotFoundException;
+import com.sunbeam.tikito.exceptions.InvalidShowException;
 import com.sunbeam.tikito.exceptions.UserNotFoundException;
 import com.sunbeam.tikito.services.BookingService;
 import com.sunbeam.tikito.entity.BookedSeatsEntity;
@@ -33,21 +37,23 @@ import jakarta.transaction.Transactional;
 
 @Transactional
 @Service
-public class BookingServiceImpl implements BookingService {
-	
+public class BookingServiceImpl implements BookingService 
+{	
 	private UserDao userDao;
 	private ShowDao showDao;
 	private SeatDao seatDao;
 	private BookedSeatsDao bookedSeatDao;
 	private BookingDao bookingDao;
+	private ModelMapper mapper;
 	
-	public BookingServiceImpl(UserDao userDao, ShowDao showDao, SeatDao seatDao, BookedSeatsDao bookedSeatDao, BookingDao bookingDao)
+	public BookingServiceImpl(UserDao userDao, ShowDao showDao, SeatDao seatDao, BookedSeatsDao bookedSeatDao, BookingDao bookingDao, ModelMapper mapper)
 	{
 		this.userDao = userDao;
 		this.showDao = showDao;
 		this.seatDao = seatDao;
 		this.bookedSeatDao = bookedSeatDao;
 		this.bookingDao = bookingDao;
+		this.mapper = mapper;
 	}
 
 	@Override
@@ -66,65 +72,73 @@ public class BookingServiceImpl implements BookingService {
 			Optional<ShowEntity> showOptional = showDao.findById(dto.getShowId());
 			if(!showOptional.isPresent())
 			{
-				throw new ShowNotFoundException("Show is not Available");
+				throw new InvalidShowException("Show is not Available");
 			}
 			else
 			{
 				ShowEntity show = showOptional.get();
 				long venueId = show.getVenue().getVenueId();
 				
-				List<SeatEntity> seats = seatDao.findBySeatIdIn(dto.getSeatIds());
-				if(seats.size() != dto.getSeatIds().size())
+				if(isDateValid(show))
 				{
-					throw new InvalidSeatsException("seats are invalid");
+					List<SeatEntity> seats = seatDao.findBySeatIdIn(dto.getSeatIds());
+					if(seats.size() != dto.getSeatIds().size())
+					{
+						throw new InvalidSeatsException("seats are invalid");
+					}
+					else
+					{
+						for(SeatEntity s : seats)
+						{
+							if(s.getVenue().getVenueId() != venueId)
+							{
+								throw new InvalidSeatsException("seats are invalid");
+							}
+						}
+						
+						List<BookedSeatsEntity> bookedSeats = bookedSeatDao.findByShowShowIdAndSeatSeatIdIn(dto.getShowId(), dto.getSeatIds());
+						if(!bookedSeats.isEmpty())
+						{
+							throw new InvalidSeatsException("seats are already booked");
+						}
+						
+						Double totalAmt = show.getPrice() * seats.size();
+						List<BookedSeatsEntity> newBookingSeats = new ArrayList<>();
+						
+	 					BookingEntity booking = new BookingEntity(null, user, show , new ArrayList<>(), totalAmt, PaymentStatus.PAID, BookingStatus.SUCCESS);
+						
+	 					for(SeatEntity s : seats)
+	 					{
+	 						BookedSeatsEntity bs = new BookedSeatsEntity(null, booking, s, show);
+	 						newBookingSeats.add(bs);
+	 					}
+	 					
+	 					try
+	 					{
+	 						booking.setBookedSeats(newBookingSeats);
+	 	 					BookingEntity newBooking = bookingDao.save(booking);
+	 	 					
+	 	 					List<String> seatNums = new ArrayList<>();
+	 	 					List<BookedSeatsEntity> newSeatsBooked = bookedSeatDao.saveAll(newBookingSeats);
+	 	 					
+	 	 					for(BookedSeatsEntity bs : newSeatsBooked)
+	 	 					{
+	 	 						seatNums.add(bs.getSeat().getSeatNo());
+	 	 					}
+	 	 					
+	 	 					ticket = new TicketBookedDto(newBooking.getBookingId(), show.getShowId(), seatNums, totalAmt, PaymentStatus.PAID, BookingStatus.SUCCESS);
+	 					}
+	 					catch(DataIntegrityViolationException e)
+	 					{
+	 						throw new InvalidSeatsException("seat are already booked");
+	 					}
+					}
 				}
 				else
 				{
-					for(SeatEntity s : seats)
-					{
-						if(s.getVenue().getVenueId() != venueId)
-						{
-							throw new InvalidSeatsException("seats are invalid");
-						}
-					}
-					
-					List<BookedSeatsEntity> bookedSeats = bookedSeatDao.findByShowShowIdAndSeatSeatIdIn(dto.getShowId(), dto.getSeatIds());
-					if(!bookedSeats.isEmpty())
-					{
-						throw new InvalidSeatsException("seats are already booked");
-					}
-					
-					Double totalAmt = show.getPrice() * seats.size();
-					List<BookedSeatsEntity> newBookingSeats = new ArrayList<>();
-					
- 					BookingEntity booking = new BookingEntity(null, user, show , new ArrayList<>(), totalAmt, PaymentStatus.PAID, BookingStatus.SUCCESS);
-					
- 					for(SeatEntity s : seats)
- 					{
- 						BookedSeatsEntity bs = new BookedSeatsEntity(null, booking, s, show);
- 						newBookingSeats.add(bs);
- 					}
- 					
- 					try
- 					{
- 						booking.setBookedSeats(newBookingSeats);
- 	 					BookingEntity newBooking = bookingDao.save(booking);
- 	 					
- 	 					List<String> seatNums = new ArrayList<>();
- 	 					List<BookedSeatsEntity> newSeatsBooked = bookedSeatDao.saveAll(newBookingSeats);
- 	 					
- 	 					for(BookedSeatsEntity bs : newSeatsBooked)
- 	 					{
- 	 						seatNums.add(bs.getSeat().getSeatNo());
- 	 					}
- 	 					
- 	 					ticket = new TicketBookedDto(newBooking.getBookingId(), show.getShowId(), seatNums, totalAmt, PaymentStatus.PAID, BookingStatus.SUCCESS);
- 					}
- 					catch(DataIntegrityViolationException e)
- 					{
- 						throw new InvalidSeatsException("seat are already booked");
- 					}
+					throw new InvalidShowException("The show has already started");
 				}
+				
 			}
 		}
 		
@@ -156,19 +170,46 @@ public class BookingServiceImpl implements BookingService {
 				}
 				else
 				{
-					booking.setBookingStatus(BookingStatus.CANCELLED);
-					booking.setPaymentStatus(PaymentStatus.REFUNDED);
-					
-					bookingDao.save(booking);
-					bookedSeatDao.deleteByBookingBookingId(bookingId);
-					
-					cancelledTicket = new CancelTicketDto(bookingId, booking.getBookingStatus(), booking.getPaymentStatus());
+					ShowEntity show = booking.getShow();
+					if(isDateValid(show))
+					{
+						booking.setBookingStatus(BookingStatus.CANCELLED);
+						booking.setPaymentStatus(PaymentStatus.REFUNDED);
+						
+						bookingDao.save(booking);
+						bookedSeatDao.deleteByBookingBookingId(bookingId);
+						
+						cancelledTicket = new CancelTicketDto(bookingId, booking.getBookingStatus(), booking.getPaymentStatus());
+					}
+					else
+					{
+						throw new InvalidShowException("The show has already started");
+					}
 				}
 			}
-
 		}
 		
 		return cancelledTicket;
+	}
+	
+	public boolean isDateValid(ShowEntity show)
+	{
+		LocalDate sDate = show.getShowDate();
+		LocalTime sTime = show.getShowStartTime(); 
+		
+		if(LocalDate.now().isBefore(sDate))
+			return true;
+		else if(LocalDate.now().isEqual(sDate))
+		{
+			if(LocalTime.now().isBefore(sTime))
+			{
+				return true;
+			}
+			else 
+				return false;
+		}
+		else
+			return false;
 	}
 
 	@Override
@@ -252,4 +293,24 @@ public class BookingServiceImpl implements BookingService {
 		return bookingList;
 	}
 
+	//work on this one for serialization issue
+	@Override
+	public List<AllBookingsDto> getAllBookingsByShow(long showId) 
+	{
+		List<AllBookingsDto> dtos = new ArrayList<>();
+		List<BookingEntity> bookings = bookingDao.findByShowShowId(showId);
+		for(BookingEntity b : bookings)
+		{
+			AllBookingsDto dto = mapper.map(b, AllBookingsDto.class);
+			dtos.add(dto);
+		}
+		
+		return dtos;
+	}
+
+//	@Override
+//	public List<String> getAllAvailableSeats(long showId) 
+//	{
+//		ShowEntity show = 
+//	}
 }
